Assistant checkpoint: Simplified sidebar with new tab system and cleaner UI
Assistant generated file changes:
- main.py: Simplify sidebar interface with tabs

---

User prompt:

I would like @main.pyto be recoded. specifically the left hand sidebar. I would like it to be simplified. 1. There should be a New tab which contains the selections we currently have so that you can create a new contour these are Location (text entry), Travel Mode (drop down) then the max distance and contour interval sldiers. finally a "Create Contour" button should be at the bottom. 2. Then for each created contour I would like a tab that contains the values entered for each of these fields as well as a "Delete Contour" button that remvoes the contoru from the map and deletes the tab. 3. When the "Create Contour" button is pressed, the new tab should be created and the API request sent which when recieved the map should be updated.