Assistant checkpoint: Implement Contour class for better state management
Assistant generated file changes:
- contour.py: Add Contour class for better state management
- main.py: Update main.py to use Contour class, Update contour creation and display logic, Update tab creation and display, Update existing contour display, Update map creation
- map_visualizer.py: Update map visualizer to use Contour class

---

User prompt:

Something strange is going on. I would like you totake an extar few seconds, re read all of the code before making small changes taht aren't fixing the underlying issues. When a new contour is created it should be an object that has its parameters stored so that they can be seen in the tab for that contour. one of these properties will be colour which will be editable and if it is eddited the contoru should be re drawn. The Contours should also be stored in this object and are populated by the API call that is made to either MapBox or Geoapify. In this way teh required information is stored where it can easily be accessed. Please make the necesary changes to implement what I have described above