Assistant checkpoint: Refactor MapVisualizer to work with Contour class
Assistant generated file changes:
- map_visualizer.py: Refactor MapVisualizer to work with Contour class

---

User prompt:

An error occurred: name 'mode' is not defined. It seems like there are alot of errors with the code not working with the new Contour class. Please re read all of the code and make all necesary changes to get the code working without these erorrs